feature: lookup intent parser
A new hashmap based intent parser.

The normalized/canonical form of an utterance serves as the key and the value is tuple of `(intent_id, [vec_of_slots_ids])`

Once a lookup is done at inference, the intent and slots are retrieved by matching their ids to a vec of intent names and a vec of slot names respectively.

This is the rust implementation of https://github.com/snipsco/snips-nlu/pull/759